Moved common test setup into init, sped up tests
