+++ hera_mc/tests/test_connections.py
@@ -11,23 +11,13 @@ record.
 
 from __future__ import absolute_import, division, print_function
 
-import unittest
 from astropy.time import Time, TimeDelta
 
 from hera_mc import part_connect, mc
+from hera_mc.tests import TestHERAMC
 
 
-class test_connections(unittest.TestCase):
-
-    def setUp(self):
-        self.test_db = mc.connect_to_mc_testing_db()
-        self.test_conn = self.test_db.engine.connect()
-        self.test_trans = self.test_conn.begin()
-        self.test_session = mc.MCSession(bind=self.test_conn)
-
-    def tearDown(self):
-        self.test_trans.rollback()
-        self.test_conn.close()
+class TestConnections(TestHERAMC):
 
     def test_add_part(self):
         self.test_session.add(part_connect.Parts())
