+++ hera_mc/tests/test_host_status.py
@@ -14,19 +14,10 @@ from __future__ import absolute_import, division, print_function
 import unittest
 
 from hera_mc import host_status, mc
+from hera_mc.tests import TestHERAMC
 
 
-class test_host_status(unittest.TestCase):
-
-    def setUp(self):
-        self.test_db = mc.connect_to_mc_testing_db()
-        self.test_conn = self.test_db.engine.connect()
-        self.test_trans = self.test_conn.begin()
-        self.test_session = mc.MCSession(bind=self.test_conn)
-
-    def tearDown(self):
-        self.test_trans.rollback()
-        self.test_conn.close()
+class TestHostStatus(TestHERAMC):
 
     def test_add_one(self):
         self.test_session.add(host_status.HostStatus())
