+++ hera_mc/tests/test_observations.py
@@ -15,20 +15,10 @@ from astropy.coordinates import EarthLocation
 from hera_mc import mc, cm_transfer, geo_handling
 from hera_mc.observations import Observation
 from hera_mc import utils
+from hera_mc.tests import TestHERAMC
 
 
-class test_hera_mc(unittest.TestCase):
-
-    def setUp(self):
-        self.test_db = mc.connect_to_mc_testing_db()
-        self.test_conn = self.test_db.engine.connect()
-        self.test_trans = self.test_conn.begin()
-        self.test_session = mc.MCSession(bind=self.test_conn)
-        cm_transfer._initialization(self.test_session)
-
-    def tearDown(self):
-        self.test_trans.rollback()
-        self.test_conn.close()
+class TestObservation(TestHERAMC):
 
     def test_new_obs(self):
         t1 = Time('2016-01-10 01:15:23', scale='utc')
