+++ hera_mc/tests/test_temperatures.py
@@ -13,19 +13,10 @@ from math import floor
 import numpy as np
 from astropy.time import Time, TimeDelta
 from hera_mc import mc, temperatures
+from hera_mc.tests import TestHERAMC
 
 
-class test_temperatures(unittest.TestCase):
-
-    def setUp(self):
-        self.test_db = mc.connect_to_mc_testing_db()
-        self.test_conn = self.test_db.engine.connect()
-        self.test_trans = self.test_conn.begin()
-        self.test_session = mc.MCSession(bind=self.test_conn)
-
-    def tearDown(self):
-        self.test_trans.rollback()
-        self.test_conn.close()
+class TestTemperatures(TestHERAMC):
 
     def test_add_paper_temps(self):
         t1 = Time('2016-01-10 01:15:23', scale='utc')
